requirements.txt
# Mobile GUI Agent Framework dependencies
anthropic>=0.18.0
google-generativeai>=0.3.0
openai>=1.12.0
pillow>=10.0.0
opencv-python>=4.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# GPT Model Implementation dependencies
torch>=2.0.0
tqdm>=4.65.0
numpy>=1.24.0
